refactor(installer): better use of GNU stow
Now using those two options:

--dotfiles
           Enable special handling for "dotfiles" (files or folders whose name begins with a period) in the package directory. If this option is enabled,
           Stow will add a preprocessing step for each file or folder whose name begins with "dot-", and replace the "dot-" prefix in the name by a period
           (.). This is useful when Stow is used to manage collections of dotfiles, to avoid having a package directory full of hidden files.

           For example, suppose we have a package containing two files, stow/dot-bashrc and stow/dot-emacs.d/init.el. With this option, Stow will create
           symlinks from .bashrc to stow/dot-bashrc and from .emacs.d/init.el to stow/dot-emacs.d/init.el. Any other files, whose name does not begin with
           "dot-", will be processed as usual.

--no-folding
           Disable folding of newly stowed directories when stowing, and refolding of newly foldable directories when unstowing.

The default action of Stow is to install a package. This means creating symlinks in the target tree that point into the package tree.  Stow attempts to do this with as few symlinks as possible; in other words, if Stow can create a single symlink that points to an entire subtree within the package tree, it will choose to do that rather than create a directory in the target tree and populate it with symlinks.
The problem with tree-folding is that extraneous files can be added on behalf of the users (e.g. hooks in git folder) and, then, make management of the repository difficult (infinite gitignore)