fix(submodule): ignore untracked (and unignored) content in the submodule
Note about the value of 'ignore":

"dirty": Only differences of the commit recorded in the superproject and the submodules HEAD will be considered modifications, all changes to the work tree of the submodule will be ignored.
When using this value, the submodule will not be scanned for work tree changes at all, leading to a performance benefit on large submodules.

"untracked": Only untracked files in the submodules work tree are ignored, a changed HEAD and/or modified files in the submodule will mark it as modified.